Place the connection matrix subsection into its own section.
I don't know what's the right thing to do here.  Either we have
a section on computing in de Rham cohomology (plus connection)
and then it's difficult to find a sensible title for the first
subsection, and the section looks like 1. huge subsection plus
2. tiny add-on.  The other way round we have one section which
is quite short.  Anyway, I thought this was slightly nicer as
it keeps the two things separate, which is nice as the reductions
in cohomology really don't relate to the connection matrix
computation at all.

Anyway, if you want to change it back, go for it :)