Change Remark 4.11.
Previous, we omitted f_5 and had an equation for f_r which was
actually wrong for f_5.  I don't think that we should do this
and have altered the remark accordingly.  I suppose the reason
you made this change originally is because the version from my
thesis was unduly lengthy (given that this is really a minor
point).  I agree, and I hope that the current version is short
enough while still being correct.