Small improvement to Prop 4.6.
For the lower bound, we actually have to use the m-pr is an
integer, so I added that back in.  Also, I removed that
m,r are non-negative from the hypotheses since we don't need
this here.  (And it is quite nice that this way, you can see
that this proposition justifies simplifyin the double sum
notation to a sum where r >= 0.  Of course, we don't have to
spell that out explicitly.)