change remark about LUP decomposition
I have recently understood that computing Delta_k^{-1} as the magma code
is doing at the moment is never a good idea, and might be worse than
solving from scratch in every instance of decompose. Since it is still
in the complexity estimates, I haven't taken it out completely. The real
practical improvement comes from sparse LUP and is enormous!